Extract utilities under the intern namespace
