+++ lib/sqlite-store.js
@@ -6,7 +6,7 @@
 var Sqlite = require('sqlite3')
 
 var _ = require('lodash')
-var Uuid = require('node-uuid')
+const { intern } = require('./intern')
 var error = require('eraro')({package: 'seneca-sqlite-store'})
 
 var name = 'sqlite-store'
@@ -152,10 +152,10 @@ module.exports = function (options) {
 
     save: function (args, done) {
       const ent = args.ent
-      const new_ent_id = ensure_id(ent)
+      const new_ent_id = intern.ensure_id(ent)
       const upsert_fields = is_upsert(args)
 
-      if (is_new(ent) && upsert_fields) {
+      if (intern.is_new(ent) && upsert_fields) {
         return do_upsert(upsert_fields, ent, new_ent_id, done)
       }
 
@@ -212,18 +212,6 @@ module.exports = function (options) {
           })
         })
       })
-
-      function ensure_id(ent) {
-        if (undefined !== ent.id$) {
-          return ent.id$
-        }
-
-        return Uuid()
-      }
-
-      function is_new(ent) {
-        return !('id' in ent)
-      }
 
       function is_upsert(args) {
         if (!Array.isArray(args.q.upsert$)) {
